requirements.txt
rpi-lgpio
pytest-mock
systemd-python

